requirements.txt
certifi==2018.8.24
django==2.2.6
djangorestframework==3.10.3